сделал вьюшки для варинатов игр
